Expose IOError and IOResult<T> in the root namespace
