+++ one_collect/src/lib.rs
@@ -5,3 +5,12 @@ pub mod perf_event;
 pub mod session;
 
 pub use sharing::{Writable, ReadOnly};
+
+pub type IOResult<T> = std::io::Result<T>;
+pub type IOError = std::io::Error;
+
+pub fn io_error(message: &str) -> IOError {
+    IOError::new(
+        std::io::ErrorKind::Other,
+        message)
+}
